requirements.txt
flask
flask-cors
tensorflow==2.12.0
opencv-python
numpy
pillow
web3
eth-account
py-solc-x
python-dotenv
werkzeug
moviepy
pycryptodome
